修复无法在 GitHub Pages 环境下查看 <eod /> 元素的问题；更改 AnEoT.Vintage.Tool 的命令行（fix-navbar => fix-github-pages）；现在当生成静态网页的目标文件夹不存在时，会自动创建这个文件夹
